refactor(filter): moves defimpl Inspect to corresponding struct files
Colocates defimpl Inspect with the corresponding struct. This should help colocate the struct itself
with the IO representation of the entity.
