+++ lib/ash/filter/filter.ex
@@ -852,3 +852,134 @@ defmodule Ash.Filter do
 
   defp add_error(%{errors: errors} = filter, error), do: %{filter | errors: [error | errors]}
 end
+
+defimpl Inspect, for: Ash.Filter do
+  import Inspect.Algebra
+  import Ash.Filter.InspectHelpers
+
+  def inspect(%Ash.Filter{impossible?: impossible, not: not_filter} = filter, opts)
+      when not is_nil(not_filter) do
+    impossible =
+      if impossible do
+        "X"
+      else
+        ""
+      end
+
+    if root?(opts) do
+      concat([
+        "#Filter<#{impossible} not (",
+        to_doc(not_filter, make_non_root(opts)),
+        ") and ",
+        to_doc(%{filter | not: nil}, make_non_root(opts)),
+        ">"
+      ])
+    else
+      concat([
+        "not (",
+        to_doc(not_filter, make_non_root(opts)),
+        ") and ",
+        to_doc(%{filter | not: nil}, make_non_root(opts))
+      ])
+    end
+  end
+
+  def inspect(
+        %Ash.Filter{
+          ors: ors,
+          relationships: relationships,
+          attributes: attributes,
+          impossible?: impossible
+        },
+        opts
+      )
+      when ors in [nil, []] and relationships in [nil, %{}] and attributes in [nil, %{}] do
+    impossible =
+      if impossible do
+        "X"
+      else
+        ""
+      end
+
+    if root?(opts) do
+      concat(["#Filter<#{impossible}", to_doc(nil, opts), ">"])
+    else
+      concat([impossible])
+    end
+  end
+
+  def inspect(%{impossible?: impossible} = filter, opts) do
+    rels =
+      filter
+      |> Map.get(:relationships)
+      |> case do
+        rels when rels == %{} ->
+          []
+
+        rels ->
+          Enum.map(rels, fn {key, value} ->
+            to_doc(value, add_to_path(opts, key))
+          end)
+      end
+
+    attrs =
+      filter
+      |> Map.get(:attributes)
+      |> case do
+        attrs when attrs == %{} ->
+          []
+
+        attrs ->
+          Enum.map(attrs, fn {key, value} ->
+            to_doc(value, put_attr(opts, key))
+          end)
+      end
+
+    and_container =
+      case attrs ++ rels do
+        [] ->
+          empty()
+
+        and_clauses ->
+          Inspect.Algebra.container_doc("(", and_clauses, ")", opts, fn term, _ -> term end,
+            break: :flex,
+            separator: " and"
+          )
+      end
+
+    all_container =
+      case Map.get(filter, :ors) do
+        nil ->
+          and_container
+
+        [] ->
+          and_container
+
+        ors ->
+          inspected_ors = Enum.map(ors, fn filter -> to_doc(filter, make_non_root(opts)) end)
+
+          Inspect.Algebra.container_doc(
+            "",
+            [and_container | inspected_ors],
+            "",
+            opts,
+            fn term, _ -> term end,
+            break: :strict,
+            separator: " or "
+          )
+      end
+
+    impossible =
+      if impossible do
+        "X"
+      else
+        ""
+      end
+
+    if root?(opts) do
+      concat(["#Filter<#{impossible}", all_container, ">"])
+    else
+      concat([impossible, all_container])
+    end
+  end
+end
